requirements.txt
discord.py==2.3.2
google-api-python-client==2.121.0
python-dotenv==1.0.1 